Add AE and AP as U.S. states.
These are foreign militrary addresses that we need for customer project. No
other SmartLogic project uses this gem, and if they did they reference the old
HEAD in Gemfile.lock.

I realize this fork is open source and upstream is still not updated for rails
4. If you are a third party relying on us, you should probably peg your Gemfile
to the parent of this commit.